refactor(drag-drop): split drag effects and add waveform tests
