+++ src/egui_app/controller/drag_drop_controller/drag_effects/mod.rs
@@ -1,0 +1,4 @@
+mod folder_moves;
+mod sample_drop;
+mod selection_drop;
+mod waveform_drop;
